+++ AlchemistCode/Cards/Ancient/Panacea.cs
@@ -1,5 +1,4 @@
 using Alchemist.AlchemistCode.Powers;
-using BaseLib.Utils;
 using MegaCrit.Sts2.Core.Commands;
 using MegaCrit.Sts2.Core.Entities.Cards;
 using MegaCrit.Sts2.Core.GameActions.Multiplayer;
@@ -10,17 +9,15 @@ namespace Alchemist.AlchemistCode.Cards.Ancient;
 [CardTheme(CardTheme.Poison, CardTheme.Antitoxin)]
 public class Panacea : AlchemistCard
 {
-    public Panacea() : base(3, CardType.Power, CardRarity.Ancient, TargetType.Self)
+    public Panacea() : base(1, CardType.Power, CardRarity.Ancient, TargetType.Self)
     {
-        WithVar("SelfPoison", 2, 0);
-        WithVar("antitoxin", 1, 1);
+        WithCostUpgradeBy(-1);
         WithTip(typeof(PoisonPower));
         WithTip(typeof(AntitoxinPower));
     }
 
     protected override async Task OnPlay(PlayerChoiceContext choiceContext, CardPlay play)
     {
-        await PowerCmd.Apply<PanaceaPower>(choiceContext, Owner.Creature,
-            DynamicVars["antitoxin"].IntValue, Owner.Creature, this);
+        await PowerCmd.Apply<PanaceaPower>(choiceContext, Owner.Creature, 1, Owner.Creature, this);
     }
 }