+++ AlchemistCode/Cards/Uncommon/Resolve.cs
@@ -1,29 +1,35 @@
-using BaseLib.Utils;
 using MegaCrit.Sts2.Core.Commands;
 using MegaCrit.Sts2.Core.Entities.Cards;
 using MegaCrit.Sts2.Core.GameActions.Multiplayer;
+using MegaCrit.Sts2.Core.Localization;
 using MegaCrit.Sts2.Core.Models.Powers;
 
 namespace Alchemist.AlchemistCode.Cards.Uncommon;
 
-[CardTheme(CardTheme.Poison)]
+[CardTheme(CardTheme.Ferment)]
 public class Resolve : AlchemistCard
 {
-    public Resolve() : base(2, CardType.Power, CardRarity.Uncommon, TargetType.Self)
+    protected override bool Ferments => true;
+
+    public Resolve() : base(1, CardType.Power, CardRarity.Uncommon, TargetType.Self)
     {
-        WithVar("Amount", 2, 1);
-        WithVar("poison", 4, 0);
+        WithVar("Amount", 1, 1);
+        WithVar("perTurn", 1, 0);
+        WithKeyword(CardKeyword.Retain);
         WithTip(typeof(StrengthPower));
-        WithTip(typeof(DexterityPower));
-        WithTip(typeof(PoisonPower));
+    }
+
+    private int Resolved => DynamicVars["Amount"].IntValue
+        + DynamicVars["perTurn"].IntValue * FermentTurns;
+
+    protected override void AddExtraArgsToDescription(LocString description)
+    {
+        base.AddExtraArgsToDescription(description);
+        description.Add("Resolved", FermentTurns > 0 ? $" ([green]{Resolved}[/green])" : "");
     }
 
     protected override async Task OnPlay(PlayerChoiceContext choiceContext, CardPlay play)
     {
-        var amount = DynamicVars["Amount"].IntValue;
-        await PowerCmd.Apply<StrengthPower>(choiceContext, Owner.Creature, amount, Owner.Creature, this);
-        await PowerCmd.Apply<DexterityPower>(choiceContext, Owner.Creature, amount, Owner.Creature, this);
-        await PowerCmd.Apply<PoisonPower>(choiceContext, Owner.Creature,
-            DynamicVars["poison"].IntValue, Owner.Creature, this);
+        await PowerCmd.Apply<StrengthPower>(choiceContext, Owner.Creature, Resolved, Owner.Creature, this);
     }
 }